refactor: 주니어 관련 화면 폴더별로 분리
