+++ lib/features/junior/presentation/views/write/junior_success_screen.dart
